requirements.txt
tushare>=1.4.21
pandas>=2.2.3
matplotlib>=3.9.2
numpy>=2.1.3
python-docx>=1.1.2
python-dotenv>=1.0.1
openpyxl>=3.1.5
